feat: remimplement txnData (#694)
* feat: remimplement txnData

* fix: debug

* fix(txnData): partly fix cumulativeGasUsed

* fix: cumulative gas in a block is sum of gas used - refund in each tx

* test(rlpAddr): too much gas

* feat(txndata): rename package from txn_data to txndata

* fix: changes in constraints

* fix: faulthy package import

---------

+++ tracer/arithmetization/src/main/java/net/consensys/linea/zktracer/module/Util.java
@@ -157,7 +157,7 @@ public class Util {
    * @param txType
    * @return transaction type
    */
-  public static int getTxTypeAsInt(TransactionType txType) {
+  public static short getTxTypeAsInt(TransactionType txType) {
     return switch (txType) {
       case FRONTIER -> 0;
       case ACCESS_LIST -> 1;
